refactor: Move countries.ts to utils/formats
